+++ lib/objects/voucher_children/limited_amount_voucher_without_end_time.dart
@@ -46,7 +46,7 @@ class LimitedAmountVoucherWithoutEndTime extends Voucher
     String? voucherID,
     String? voucherName,
     DateTime? startTime,
-    double? discountValue,
+    int? discountValue,
     int? minimumPurchase,
     int? maxUsagePerPerson,
     DistributionType? distributionType,
